requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.2.10
asn1crypto==1.4.0
attrs==19.3.0
autopep8==1.5.4
bcrypt==3.1.7
black==19.10b0
blinker==1.4
Brlapi==0.7.0
capstone==4.0.2
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
command-not-found==0.3
crypto==1.4.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1
djangorestframework==3.11.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
httplib2==0.14.0
idna==2.8
intervaltree==3.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
M2Crypto==0.31.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
Naked==0.1.31
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.4
paramiko==2.6.0
pathspec==0.8.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.7.2
pwntools==4.2.1
pycairo==1.16.2
pycodestyle==2.6.0
pycrypto==2.6.1
pycryptodomex==3.9.8
pycups==1.9.73
pyelftools==0.26
Pygments==2.6.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.4
PySocks==1.7.1
python-apt==2.0.0
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2020.7.14
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
ROPGadget==6.3
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
six==1.14.0
sortedcontainers==2.2.2
sqlparse==0.3.1
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
toml==0.10.1
typed-ast==1.4.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc3
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.21
wadllib==1.3.3
xkit==0.0.0
